+++ src/variables/ManifestVariable.php
@@ -30,10 +30,10 @@ class ManifestVariable
      * @param bool       $async
      * @param null|array $config
      *
-     * @return \Twig_Markup
+     * @return \Twig\Markup
      * @throws \yii\web\NotFoundHttpException
      */
-    public function includeCssModule(string $moduleName, bool $async = false, $config = null): \Twig_Markup
+    public function includeCssModule(string $moduleName, bool $async = false, $config = null): \Twig\Markup
     {
         return Template::raw(
             Twigpack::$plugin->manifest->getCssModuleTags($moduleName, $async, $config)
@@ -47,7 +47,7 @@ class ManifestVariable
      *
      * @return string
      */
-    public function includeInlineCssTags(string $path): \Twig_Markup
+    public function includeInlineCssTags(string $path): \Twig\Markup
     {
         return Template::raw(
             Twigpack::$plugin->manifest->getCssInlineTags($path)
@@ -60,9 +60,9 @@ class ManifestVariable
      * @param null|string $name
      * @param null|array $config
      *
-     * @return \Twig_Markup
+     * @return \Twig\Markup
      */
-    public function includeCriticalCssTags($name = null, $config = null): \Twig_Markup
+    public function includeCriticalCssTags($name = null, $config = null): \Twig\Markup
     {
         return Template::raw(
             Twigpack::$plugin->manifest->getCriticalCssTags($name, $config)
@@ -73,9 +73,9 @@ class ManifestVariable
      * Returns the uglified loadCSS rel=preload Polyfill as per:
      * https://github.com/filamentgroup/loadCSS#how-to-use-loadcss-recommended-example
      *
-     * @return \Twig_Markup
+     * @return \Twig\Markup
      */
-    public static function includeCssRelPreloadPolyfill(): \Twig_Markup
+    public static function includeCssRelPreloadPolyfill(): \Twig\Markup
     {
         return Template::raw(
             Twigpack::$plugin->manifest->getCssRelPreloadPolyfill()
@@ -87,7 +87,7 @@ class ManifestVariable
      * @param bool       $async
      * @param null|array $config
      *
-     * @return null|\Twig_Markup
+     * @return null|\Twig\Markup
      * @throws \yii\web\NotFoundHttpException
      */
     public function includeJsModule(string $moduleName, bool $async = false, $config = null)
@@ -104,7 +104,7 @@ class ManifestVariable
      * @param string $type
      * @param null   $config
      *
-     * @return null|\Twig_Markup
+     * @return null|\Twig\Markup
      * @throws \yii\web\NotFoundHttpException
      */
     public function getModuleUri(string $moduleName, string $type = 'modern', $config = null)
@@ -117,9 +117,9 @@ class ManifestVariable
     /**
      * Include the Safari 10.1 nomodule fix JavaScript
      *
-     * @return \Twig_Markup
+     * @return \Twig\Markup
      */
-    public function includeSafariNomoduleFix(): \Twig_Markup
+    public function includeSafariNomoduleFix(): \Twig\Markup
     {
         return Template::raw(
             Twigpack::$plugin->manifest->getSafariNomoduleFix()
@@ -131,9 +131,9 @@ class ManifestVariable
      *
      * @param string $path
      *
-     * @return \Twig_Markup
+     * @return \Twig\Markup
      */
-    public function includeFile(string $path): \Twig_Markup
+    public function includeFile(string $path): \Twig\Markup
     {
         return Template::raw(
             Twigpack::$plugin->manifest->getFile($path)
@@ -147,9 +147,9 @@ class ManifestVariable
      * @param string $type
      * @param null   $config
      *
-     * @return \Twig_Markup
+     * @return \Twig\Markup
      */
-    public function includeFileFromManifest(string $fileName, string $type = 'legacy', $config = null): \Twig_Markup
+    public function includeFileFromManifest(string $fileName, string $type = 'legacy', $config = null): \Twig\Markup
     {
         return Template::raw(
             Twigpack::$plugin->manifest->getFileFromManifest($fileName, $type, $config)